+++ lib/kinetic_player.dart
@@ -13,3 +13,4 @@ export 'src/gsy/gsy_ui_config.dart';
 export 'src/gsy/gsy_video_features.dart';
 export 'src/gsy/gsy_video_controller_impl.dart';
 export 'src/sg/sg_video_controller_impl.dart';
+export 'src/sg/sg_video_features.dart';